cli: split report UI app.rs and extract BalanceView
app.rs had grown past 2200 lines, mixing the search machinery, the
error/overlay popup, the balance and register row/view types, the
snapshot types, and the whole App state machine in one file.

Split the self-contained type clusters into sibling modules — search,
overlay, register, balance — and group the balance-screen state (tree,
mode, fold set, rows, nav, search) into a BalanceView struct mirroring
the existing RegisterView. App is now a thin dispatcher: update()
delegates balance transitions to BalanceView and owns only the
cross-screen concerns (focused screen, overlay, footer notice, reload
snapshot). Behavior is unchanged; the single-Message/single-update TEA
shape is preserved.

Refs #425.
